+++ packages/react-dom/src/client/validateDOMNesting.js
@@ -7,7 +7,7 @@
 
 import warning from 'shared/warning';
 // TODO: direct imports like some-package/src/* are bad. Fix me.
-import {getCurrentFiberStackInDevOrNull} from 'react-reconciler/src/ReactDebugCurrentFiber';
+import {getCurrentFiberStackInDevOrNull} from 'react-reconciler/src/ReactCurrentFiber';
 
 let validateDOMNesting = () => {};
