+++ packages/react-reconciler/src/ReactCurrentFiber.js
